refactored dir names
